WIP Modpack v1.5
This is the README for the Work In Progress ModPack v1.5

Because I suffered from multiple savegame corruptions and the Mutant Loot mod author claiming it to be user-error.
I decided to remove all item-mods.
It may very well be, a game-side bug, that causes these issues. But claiming it to be user error, and insisting that nothing is wrong, just doesn't sit well with me.

For my part, I am done with item-mods for now.